+++ WeightInsight/Models/StatisticDataObject.swift
@@ -11,8 +11,8 @@ import RealmSwift
 class StatisticDataObject: Object, Identifiable {
     @objc dynamic var id = Date().formattedString(format: "yyyy-MM-dd")
     @objc dynamic var weight: Double = 0
-    @objc dynamic var steps: Int = 0
-    @objc dynamic var calories: Int = 0
+    @objc dynamic var steps: Double = 0
+    @objc dynamic var calories: Double = 0
     @objc dynamic var date: Date = Date()
     
     override static func primaryKey() -> String? {
